Error hendling

+++ src/api/doctor_card/doctor_card.service.ts
@@ -1,9 +1,14 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateDoctorCardDto } from './dto/create-doctor_card.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { successRes } from 'src/infrostructure/utils/succesResponse';
 import { ErrorHender } from 'src/infrostructure/utils/catchError';
 import { Request } from 'express';
+import { ERols } from 'src/common/enum';
 
 @Injectable()
 export class DoctorCardService {
@@ -21,11 +26,14 @@ export class DoctorCardService {
     }
   }
 
-  async findAll() {
+  async findAll(req: Request) {
+    const user = req['user'];
     try {
-      const data = await this.prisma.doctor_card.findMany();
+      const where = user.role === ERols.DOCTOR ? { doctor_id: user.id } : {};
+
+      const data = await this.prisma.doctor_card.findMany({ where });
       if (!data.length) {
-        throw new NotFoundException();
+        throw new NotFoundException('Doctor kartalari topilmadi');
       }
       return successRes(data);
     } catch (error) {
@@ -33,13 +41,19 @@ export class DoctorCardService {
     }
   }
 
-  async findOne(id: number) {
+  async findOne(id: number, req: Request) {
+    const user = req['user'];
     try {
       const data = await this.prisma.doctor_card.findUnique({
         where: { id },
       });
       if (!data) {
-        throw new NotFoundException();
+        throw new NotFoundException('Doctor kartasi topilmadi');
+      }
+      const isOwner = data.doctor_id === user.id;
+      const isAdmin = [ERols.ADMIN, ERols.SUPPER_ADMIN].includes(user.role);
+      if (!isOwner && !isAdmin) {
+        throw new ForbiddenException('Siz bu kartaga kira olmaysiz');
       }
       return successRes(data);
     } catch (error) {
@@ -47,15 +61,22 @@ export class DoctorCardService {
     }
   }
 
-  async remove(id: number) {
+  async remove(id: number, req: Request) {
+    const user = req['user'];
     try {
       const data = await this.prisma.doctor_card.findUnique({
         where: { id },
       });
       if (!data) {
-        throw new NotFoundException();
+        throw new NotFoundException('Doctor kartasi topilmadi');
+      }
+      const isOwner = data.doctor_id === user.id;
+      const isAdmin = [ERols.ADMIN, ERols.SUPPER_ADMIN].includes(user.role);
+      if (!isOwner && !isAdmin) {
+        throw new ForbiddenException("Siz bu kartani o'chira olmaysiz");
       }
       await this.prisma.doctor_card.delete({ where: { id } });
+
       return { message: 'Deleted', statusCode: 200 };
     } catch (error) {
       return ErrorHender(error);
